feat: apply context API
