+++ src/App.jsx
@@ -3,16 +3,19 @@ import Routers from "../src/routers/Routers";
 import Header from "./components/Header/Header";
 import Footer from "./components/Footer/Footer";
 import React from "react";
+import { GetListProvider } from "./context/getCarList";
 
 function App() {
   return (
-    <div className="App">
-      <Header />
-      <div>
-        <Routers />
+    <GetListProvider>
+      <div className="App">
+        <Header />
+        <div>
+          <Routers />
+        </div>
+        <Footer />
       </div>
-      <Footer />
-    </div>
+    </GetListProvider>
   );
 }
